+++ src/wam2layers/cli.py
@@ -11,9 +11,11 @@ import logging
 import shutil
 from datetime import datetime
 from pathlib import Path
+import sys
 
 import click
 
+from wam2layers import __version__
 from wam2layers.analysis import visualization
 from wam2layers.config import Config
 from wam2layers.preprocessing.era5 import prep_experiment
@@ -51,7 +53,17 @@ def _copy_config_yaml(yaml_path, target_path):
     shutil.copy(yaml_path, target_path)
 
 
+def get_wam_version():
+    """Get the version of WAM2layers"""
+    pkg_dir = Path(__file__).parent.absolute()
+    return (
+        f"wam2layers {__version__} from {pkg_dir} "
+        f"(python {sys.version_info.major}.{sys.version_info.minor})"
+    )
+
+
 @click.group()
+@click.version_option(message=get_wam_version())
 def cli():
     """Command line interface to WAM2layers."""
     pass